graph: Stop walking if we have already seen this path going in this direction before
`next :abort` causes the tree traversal to stop going this direction.
Once we know something else has gone down this path, we don't need to traverse anymore.

This makes this function runtime be linear with size of the map instead of exponential.

Using this as a test case:
Before:
Graph computed with 3088734 completed walk calls (skipped 586795) and found 93 hexes, 56 nodes, 252 paths

After:
 Graph computed with 34996 completed walk calls (skipped 6511) and found 93 hexes, 56 nodes, 252 paths

This is for https://github.com/tobymao/18xx/issues/10331